[FIX] add static cast to fix return type

+++ src/main/cpp/MSNumpress.cpp
@@ -508,7 +508,7 @@ int decodeSafe(
 		return -1;
 	}
 	
-	return ri;
+	return static_cast<int>(ri);
 }
 
 /////////////////////////////////////////////////////////////
